+++ spinta/cli/manifest.py
@@ -1,20 +1,25 @@
 import pathlib
 from typing import Iterator
 from typing import List
+from typing import Optional
 
 from typer import Argument
 from typer import Context as TyperContext
 from typer import Option
 from typer import Typer
+from typer import echo
 
 from spinta import commands
 from spinta.components import Context
 from spinta.core.enums import Access
 from spinta.manifests.helpers import load_manifest_nodes
-from spinta.naming.helpers import reformat_names
+from spinta.manifests.tabular.components import ManifestRow
 from spinta.manifests.tabular.helpers import datasets_to_tabular
+from spinta.manifests.tabular.helpers import normalizes_columns
 from spinta.manifests.tabular.helpers import read_tabular_manifest
+from spinta.manifests.tabular.helpers import render_tabular_manifest_rows
 from spinta.manifests.tabular.helpers import write_tabular_manifest
+from spinta.naming.helpers import reformat_names
 from spinta.utils.enums import get_enum_by_name
 from spinta.utils.imports import importstr
 
@@ -35,16 +40,21 @@ def copy(
     format_names: bool = Option(False, help=(
         "Reformat model and property names."
     )),
+    output: Optional[pathlib.Path] = Option(None, '-o', '--output', help=(
+        "Output tabular manifest in a specified file"
+    )),
+    columns: Optional[str] = Option(None, '-c', '--columns', help=(
+        "Comma separated list of columns"
+    )),
     files: List[pathlib.Path] = Argument(None, help=(
         "Source manifest files to copy from"
-    )),
-    dest: pathlib.Path = Argument(None, help=(
-        "Target manifest file to save a copy to"
     )),
 ):
     """Copy models from CSV manifest files into another CSV manifest file"""
     context: Context = ctx.obj
     access = get_enum_by_name(Access, access)
+    cols = normalizes_columns(columns.split(',')) if columns else None
+
     rows = _read_csv_files(
         context,
         files,
@@ -52,8 +62,12 @@ def copy(
         access=access,
         format_names=format_names,
     )
-    with dest.open('w') as f:
-        write_tabular_manifest(f, rows)
+
+    if output:
+        with output.open('w') as f:
+            write_tabular_manifest(f, rows)
+    else:
+        echo(render_tabular_manifest_rows(rows, cols))
 
 
 def _read_csv_files(
@@ -63,7 +77,7 @@ def _read_csv_files(
     external: bool = True,
     access: Access = Access.private,
     format_names: bool = False,
-) -> Iterator[dict]:
+) -> Iterator[ManifestRow]:
     rc = context.get('rc')
     for path in files:
         with context: